restructured NucelotideEntropy, plot utilities and entropy calc into seperate components
